Attempted bugfix.

+++ Buffers/base_buffer.py
@@ -263,7 +263,7 @@ class BaseBuffer:
                                    assay_group)
         except:
             try:
-                self.create_data_group("prey_orientations", np.array(self.prey_orientation_buffer), assay_group)
+                self.create_data_group("prey_orientations", np.array(self.prey_orientation_buffer).astype(np.float64), assay_group)
             except TypeError:
                 pass
                 # self.create_data_group("prey_orientations", np.array(self.prey_orientation_buffer).astype(np.float64), assay_group)
@@ -273,7 +273,7 @@ class BaseBuffer:
                                    assay_group)
         except:
             try:
-                self.create_data_group("predator_orientation", np.array(self.predator_orientation_buffer), assay_group)
+                self.create_data_group("predator_orientation", np.array(self.predator_orientation_buffer).astype(np.float64), assay_group)
             except TypeError:
                 pass
 
@@ -281,7 +281,7 @@ class BaseBuffer:
             self.create_data_group("prey_ages", self.pad_buffer(np.array(self.prey_age_buffer)), assay_group)
         except:
             try:
-                self.create_data_group("prey_ages", np.array(self.prey_age_buffer), assay_group)
+                self.create_data_group("prey_ages", np.array(self.prey_age_buffer).astype(np.float64), assay_group)
             except TypeError:
                 pass
 
@@ -289,7 +289,7 @@ class BaseBuffer:
             self.create_data_group("prey_gaits", self.pad_buffer(np.array(self.prey_gait_buffer)), assay_group)
         except:
             try:
-                self.create_data_group("prey_gaits", np.array(self.prey_gait_buffer), assay_group)
+                self.create_data_group("prey_gaits", np.array(self.prey_gait_buffer).astype(np.float64), assay_group)
             except TypeError:
                 pass
